+++ specparam/tests/objs/test_components.py
@@ -1,0 +1,13 @@
+"""Tests for specparam.objs.components."""
+
+from specparam.objs.components import *
+
+###################################################################################################
+###################################################################################################
+
+## ModelComponents object
+
+def test_model_components():
+
+    mc = ModelComponents()
+    assert mc
